+++ src/usuarios/Secretario.java
@@ -1,6 +1,5 @@
 package usuarios;
 
-import utilidades.LecturaEscrituraFichero;
 import java.util.ArrayList;
 
 /**
@@ -29,7 +28,7 @@ public final class Secretario extends Trabajador
 
     public Secretario(String nombre, String apellido1, String apellido2, String dni,
                       String user, String password, String email, String iban, String tipopersona,
-                      double salario) {
+                      int salario) {
         super(nombre, apellido1, apellido2, dni,  user, password, email, iban, tipopersona, salario);
     }
